Implemented the 'CreateProjMatrix()' method

+++ Arctic-Engine/source/render/CameraConfig.h
@@ -4,7 +4,7 @@ struct CameraConfig
 {
 	int width;
 	int height;
-	int fov;
+	float fov;
 	float nearClip;
 	float farClip;
 };
